requirements.txt
Django
mysqlclient
sympy
requests
google-cloud-compute